Remove redundant casts.

+++ src/main/java/jp/ac/nii/RankCalculationReducer.java
@@ -72,7 +72,7 @@ public class RankCalculationReducer extends
 	public class ValueComparator implements Comparator<Map.Entry<String,Integer>> {
 		@Override
 		public int compare(Entry<String,Integer> entry1, Entry<String,Integer> entry2) {
-			return ((Integer)entry2.getValue()).compareTo((Integer)entry1.getValue());
+			return entry2.getValue().compareTo(entry1.getValue());
 		}
 	}
 }
